feat: フォルダ作成ボタンを新しいレイアウトコンポーネントに移動し、ワークスペースメニューに統合
