+++ app/(protected)/[workspaceId]/page.tsx
@@ -1,5 +1,5 @@
 import { getFolderById } from '@/features/folder/api';
-import { CreateItemButton } from '@/features/folder/components';
+import { CreateItemButton } from '@/features/layout/components/workspace-menu/create-item-button';
 import { notFound } from 'next/navigation';
 
 type WorkspacePageProps = {
