--- features/folder/components/index.ts
@@ -1,2 +1,0 @@
-export * from './create-workspace-button';
-export * from './create-item-button';
