+++ features/layout/components/workspace-menu/folder-item.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useState } from 'react';
-import { ChevronRight, LayoutDashboard } from 'lucide-react';
+import { ChevronRight, LayoutDashboard, Folder } from 'lucide-react';
 import Link from 'next/link';
 import {
   SidebarMenuButton,
@@ -10,6 +10,7 @@ import {
   SidebarMenuSubButton,
   SidebarMenuSubItem,
 } from '@/components/ui/sidebar';
+import { CreateItemButton } from './create-item-button';
 import type { Folder as FolderType } from '@/features/folder/types';
 
 type FolderItemProps = {
@@ -45,6 +46,7 @@ export function FolderItem({ folder, level = 0 }: FolderItemProps) {
             <Link href={`/${folder.id}`} className="flex-1">
               <span>{folder.name}</span>
             </Link>
+            <CreateItemButton workspaceId={folder.id} />
           </div>
         </SidebarMenuButton>
         {hasChildren && isOpen && (
@@ -71,7 +73,7 @@ export function FolderItem({ folder, level = 0 }: FolderItemProps) {
             }}
             className="flex items-center justify-center shrink-0 rounded hover:bg-primary/10 cursor-pointer"
           >
-            <LayoutDashboard className="size-4 group-hover/item:hidden" />
+            <Folder className="size-4 group-hover/item:hidden" />
             <ChevronRight
               className={`size-4 hidden group-hover/item:block transition-transform duration-200 ${
                 isOpen ? 'rotate-90' : ''
@@ -81,6 +83,7 @@ export function FolderItem({ folder, level = 0 }: FolderItemProps) {
           <Link href={`/${folder.id}`} className="flex-1">
             <span>{folder.name}</span>
           </Link>
+          <CreateItemButton workspaceId={folder.id} />
         </div>
       </SidebarMenuSubButton>
       {hasChildren && isOpen && (
