+++ features/layout/components/workspace-menu/index.tsx
@@ -7,7 +7,7 @@ import {
 } from '@/components/ui/sidebar';
 
 import { WorkspaceItems } from './items';
-import { CreateWorkspaceButton } from '@/features/folder/components';
+import { CreateItemButton } from './create-item-button';
 import type { Folder } from '@/features/folder/types';
 
 type WorkspaceMenuProps = {
@@ -17,8 +17,10 @@ type WorkspaceMenuProps = {
 export function WorkspaceMenu({ folders }: WorkspaceMenuProps) {
   return (
     <SidebarGroup>
-      <SidebarGroupLabel>ワークスペース</SidebarGroupLabel>
-      <CreateWorkspaceButton />
+      <SidebarGroupLabel className="group/workspace hover:bg-sidebar-accent hover:text-sidebar-accent-foreground transition-colors cursor-pointer">
+        ワークスペース
+        <CreateItemButton workspaceId="" variant="workspace" />
+      </SidebarGroupLabel>
       <SidebarMenu>
         <WorkspaceItems folders={folders} />
       </SidebarMenu>
